EES-7042 - [WIP-2] - Fixing Robot Tests
