+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Services/UserManagementServicePermissionTests.cs
@@ -186,6 +186,7 @@ public class UserManagementServicePermissionTests
         UsersAndRolesDbContext? usersAndRolesDbContext = null,
         IPersistenceHelper<UsersAndRolesDbContext>? usersAndRolesPersistenceHelper = null,
         IUserRoleService? userRoleService = null,
+        IGlobalRoleService? globalRoleService = null,
         IUserRepository? userRepository = null,
         IUserService? userService = null,
         IUserPreReleaseRoleRepository? userPreReleaseRoleRepository = null,
@@ -203,6 +204,7 @@ public class UserManagementServicePermissionTests
             contentDbContext,
             usersAndRolesPersistenceHelper ?? new PersistenceHelper<UsersAndRolesDbContext>(usersAndRolesDbContext),
             userRoleService ?? Mock.Of<IUserRoleService>(Strict),
+            globalRoleService: Mock.Of<IGlobalRoleService>(Strict),
             userRepository ?? Mock.Of<IUserRepository>(Strict),
             userService ?? AlwaysTrueUserService(CreatedById).Object,
             userPreReleaseRoleRepository ?? Mock.Of<IUserPreReleaseRoleRepository>(Strict),
