+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Services/UserRoleServiceTests.cs
@@ -53,379 +53,10 @@ public abstract class UserRoleServiceTests
         fixture => fixture.DefaultSoftDeletedUser(),
     ];
 
-    public class SetGlobalRoleForUserTests : UserRoleServiceTests
-    {
-        [Fact]
-        public async Task Success()
-        {
-            var user = new ApplicationUser { Id = Guid.NewGuid().ToString() };
-
-            var role = new IdentityRole { Id = Guid.NewGuid().ToString(), Name = "Role" };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddAsync(user);
-                await userAndRolesDbContext.AddAsync(role);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            userManager.Setup(mock => mock.GetRolesAsync(ItIsUser(user))).ReturnsAsync(new List<string>());
-
-            userManager
-                .Setup(mock => mock.AddToRoleAsync(ItIsUser(user), role.Name))
-                .ReturnsAsync(new IdentityResult());
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object
-                );
-
-                var result = await service.SetGlobalRoleForUser(user.Id, role.Id);
-
-                VerifyAllMocks(userManager);
-
-                result.AssertRight();
-
-                userManager.Verify(
-                    mock =>
-                        mock.AddToRoleAsync(
-                            It.Is<ApplicationUser>(applicationUser => applicationUser.Id == user.Id),
-                            role.Name
-                        ),
-                    Times.Once
-                );
-            }
-        }
-
-        [Fact]
-        public async Task AlreadyHasSameRole()
-        {
-            var user = new ApplicationUser { Id = Guid.NewGuid().ToString() };
-
-            var role = new IdentityRole { Id = Guid.NewGuid().ToString(), Name = "Role" };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddAsync(user);
-                await userAndRolesDbContext.AddAsync(role);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            userManager.Setup(mock => mock.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(role.Name));
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object
-                );
-
-                var result = await service.SetGlobalRoleForUser(user.Id, role.Id);
-
-                VerifyAllMocks(userManager);
-
-                result.AssertRight();
-
-                VerifyAllMocks(userManager);
-            }
-        }
-
-        [Fact]
-        public async Task AlreadyHasAnotherRole()
-        {
-            var user = new ApplicationUser { Id = Guid.NewGuid().ToString() };
-
-            var newRole = new IdentityRole { Id = Guid.NewGuid().ToString(), Name = "New Role" };
-
-            var existingRole = new IdentityRole { Id = Guid.NewGuid().ToString(), Name = "Existing Role" };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddAsync(user);
-                await userAndRolesDbContext.AddRangeAsync(newRole, existingRole);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // Here we are setting up the user so that they have a different Global Role currently assigned
-            // than the new one being set.  They will have the existing one removed and the new one added.
-            userManager.Setup(mock => mock.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(existingRole.Name));
-
-            userManager
-                .Setup(mock => mock.AddToRoleAsync(ItIsUser(user), newRole.Name))
-                .ReturnsAsync(new IdentityResult());
-
-            userManager
-                .Setup(mock =>
-                    mock.RemoveFromRolesAsync(ItIsUser(user), ItIs.ListSequenceEqualTo(ListOf(existingRole.Name)))
-                )
-                .ReturnsAsync(new IdentityResult());
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object
-                );
-
-                var result = await service.SetGlobalRoleForUser(user.Id, newRole.Id);
-
-                VerifyAllMocks(userManager);
-
-                result.AssertRight();
-
-                VerifyAllMocks(userManager);
-            }
-        }
-
-        [Fact]
-        public async Task NoRole_ReturnsNotFound()
-        {
-            var user = new ApplicationUser { Id = Guid.NewGuid().ToString() };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddAsync(user);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object
-                );
-
-                var result = await service.SetGlobalRoleForUser(user.Id, Guid.NewGuid().ToString());
-
-                result.AssertNotFound();
-            }
-
-            userManager.VerifySet(mock => mock.Logger = null, Times.Once());
-            VerifyAllMocks(userManager);
-        }
-
-        [Fact]
-        public async Task NoUser_ReturnsNotFound()
-        {
-            var role = new IdentityRole { Id = Guid.NewGuid().ToString() };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddAsync(role);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object
-                );
-
-                var result = await service.SetGlobalRoleForUser(Guid.NewGuid().ToString(), role.Id);
-
-                result.AssertNotFound();
-            }
-
-            VerifyAllMocks(userManager);
-        }
-    }
-
     public class AddPublicationRoleTests : UserRoleServiceTests
     {
         [Fact]
         public async Task AddPublicationRole()
-        {
-            var user = new ApplicationUser { Id = _user.Id.ToString(), Email = _user.Email };
-
-            var publication = new Publication();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddAsync(user);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                await contentDbContext.Publications.AddAsync(publication);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            UserPublicationRole createdUserPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(_dataFixture.DefaultUser())
-                .WithPublication(_dataFixture.DefaultPublication());
-
-            var userManager = MockUserManager();
-
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(RoleNames.Analyst));
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository
-                .Setup(mock =>
-                    mock.UserHasRoleOnPublication(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Approver,
-                        ResourceRoleFilter.ActiveOnly,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(false);
-            userPublicationRoleRepository
-                .Setup(s =>
-                    s.Create(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Approver,
-                        _user.Id,
-                        null,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(createdUserPublicationRole);
-
-            var userResourceRoleNotificationService = new Mock<IUserResourceRoleNotificationService>(Strict);
-            userResourceRoleNotificationService
-                .Setup(mock =>
-                    mock.NotifyUserOfNewPublicationRole(createdUserPublicationRole.Id, It.IsAny<CancellationToken>())
-                )
-                .Returns(Task.CompletedTask);
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    contentDbContext: contentDbContext,
-                    userResourceRoleNotificationService: userResourceRoleNotificationService.Object,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object
-                );
-
-                var result = await service.AddPublicationRole(_user.Id, publication.Id, PublicationRole.Approver);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userResourceRoleNotificationService, userManager, userPublicationRoleRepository);
-        }
-
-        [Fact]
-        public async Task AddPublicationRole_HasNoRoles_UpgradesRole()
-        {
-            var user = new ApplicationUser { Id = _user.Id.ToString(), Email = _user.Email };
-
-            var publication = new Publication();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddAsync(user);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                await contentDbContext.Publications.AddAsync(publication);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            UserPublicationRole createdUserPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(_dataFixture.DefaultUser())
-                .WithPublication(_dataFixture.DefaultPublication());
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, []);
-            userPublicationRoleRepository
-                .Setup(mock =>
-                    mock.UserHasRoleOnPublication(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Drafter,
-                        ResourceRoleFilter.ActiveOnly,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(false);
-            userPublicationRoleRepository
-                .Setup(s =>
-                    s.Create(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Drafter,
-                        _user.Id,
-                        null,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(createdUserPublicationRole);
-
-            var userResourceRoleNotificationService = new Mock<IUserResourceRoleNotificationService>(Strict);
-            userResourceRoleNotificationService
-                .Setup(mock =>
-                    mock.NotifyUserOfNewPublicationRole(createdUserPublicationRole.Id, It.IsAny<CancellationToken>())
-                )
-                .Returns(Task.CompletedTask);
-
-            var userManager = MockUserManager();
-
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(user))).ReturnsAsync([]);
-
-            userManager
-                .Setup(s => s.AddToRoleAsync(ItIsUser(user), RoleNames.Analyst))
-                .ReturnsAsync(new IdentityResult());
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    contentDbContext: contentDbContext,
-                    userResourceRoleNotificationService: userResourceRoleNotificationService.Object,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object
-                );
-
-                var result = await service.AddPublicationRole(_user.Id, publication.Id, PublicationRole.Drafter);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userResourceRoleNotificationService, userManager);
-        }
-
-        [Fact]
-        public async Task AddPublicationRole_HasHigherGlobalRoleThanAnalyst_DoesNotUpgradeRole()
         {
             var user = new ApplicationUser { Id = _user.Id.ToString(), Email = _user.Email };
 
@@ -451,14 +82,18 @@ public abstract class UserRoleServiceTests
                 .WithUser(_dataFixture.DefaultUser())
                 .WithPublication(_dataFixture.DefaultPublication());
 
+            var globalRoleService = new Mock<IGlobalRoleService>(Strict);
+            globalRoleService
+                .Setup(mock => mock.UpgradeToGlobalRoleIfRequired(ItIsUser(user), RoleNames.Analyst))
+                .Returns(Task.CompletedTask);
+
             var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, []);
             userPublicationRoleRepository
                 .Setup(mock =>
                     mock.UserHasRoleOnPublication(
                         _user.Id,
                         publication.Id,
-                        PublicationRole.Drafter,
+                        PublicationRole.Approver,
                         ResourceRoleFilter.ActiveOnly,
                         It.IsAny<CancellationToken>()
                     )
@@ -469,7 +104,7 @@ public abstract class UserRoleServiceTests
                     s.Create(
                         _user.Id,
                         publication.Id,
-                        PublicationRole.Drafter,
+                        PublicationRole.Approver,
                         _user.Id,
                         null,
                         It.IsAny<CancellationToken>()
@@ -484,12 +119,6 @@ public abstract class UserRoleServiceTests
                 )
                 .Returns(Task.CompletedTask);
 
-            var userManager = MockUserManager();
-
-            // Here we are testing that if the user already has a higher-powered role than Analyst, there's no need
-            // to assign them the lower-powered Analyst role.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(RoleNames.BauUser));
-
             await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
             await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
             {
@@ -497,121 +126,16 @@ public abstract class UserRoleServiceTests
                     usersAndRolesDbContext: userAndRolesDbContext,
                     contentDbContext: contentDbContext,
                     userResourceRoleNotificationService: userResourceRoleNotificationService.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userManager: userManager.Object
-                );
-
-                var result = await service.AddPublicationRole(_user.Id, publication.Id, PublicationRole.Drafter);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userManager, userResourceRoleNotificationService);
-
-            userManager.Verify(
-                mock => mock.AddToRoleAsync(It.IsAny<ApplicationUser>(), It.IsAny<string>()),
-                Times.Never
-            );
-
-            userManager.Verify(
-                mock => mock.RemoveFromRolesAsync(It.IsAny<ApplicationUser>(), It.IsAny<List<string>>()),
-                Times.Never
-            );
-        }
-
-        [Fact]
-        public async Task AddPublicationRole_HasLowerGlobalRoleThanAnalyst_UpgradesRole()
-        {
-            var user = new ApplicationUser { Id = _user.Id.ToString(), Email = _user.Email };
-
-            var publication = new Publication();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddAsync(user);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                await contentDbContext.Publications.AddAsync(publication);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            UserPublicationRole createdUserPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(_dataFixture.DefaultUser())
-                .WithPublication(_dataFixture.DefaultPublication());
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, []);
-            userPublicationRoleRepository
-                .Setup(mock =>
-                    mock.UserHasRoleOnPublication(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Drafter,
-                        ResourceRoleFilter.ActiveOnly,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(false);
-            userPublicationRoleRepository
-                .Setup(s =>
-                    s.Create(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Drafter,
-                        _user.Id,
-                        null,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(createdUserPublicationRole);
-
-            var userResourceRoleNotificationService = new Mock<IUserResourceRoleNotificationService>(Strict);
-            userResourceRoleNotificationService
-                .Setup(mock =>
-                    mock.NotifyUserOfNewPublicationRole(createdUserPublicationRole.Id, It.IsAny<CancellationToken>())
-                )
-                .Returns(Task.CompletedTask);
-
-            var userManager = MockUserManager();
-
-            // Here we are checking to see that the user is "upgraded" to the higher-powered Analyst role by their
-            // existing PreReleaseUser role being removed and the higher-powered Analyst role being added.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(RoleNames.PrereleaseUser));
-
-            userManager
-                .Setup(s => s.AddToRoleAsync(ItIsUser(user), RoleNames.Analyst))
-                .ReturnsAsync(new IdentityResult());
-
-            userManager
-                .Setup(s =>
-                    s.RemoveFromRolesAsync(ItIsUser(user), ItIs.ListSequenceEqualTo(ListOf(RoleNames.PrereleaseUser)))
-                )
-                .ReturnsAsync(new IdentityResult());
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    contentDbContext: contentDbContext,
-                    userResourceRoleNotificationService: userResourceRoleNotificationService.Object,
-                    userManager: userManager.Object,
+                    globalRoleService: globalRoleService.Object,
                     userPublicationRoleRepository: userPublicationRoleRepository.Object
                 );
 
-                var result = await service.AddPublicationRole(_user.Id, publication.Id, PublicationRole.Drafter);
+                var result = await service.AddPublicationRole(_user.Id, publication.Id, PublicationRole.Approver);
 
                 result.AssertRight();
             }
 
-            VerifyAllMocks(userResourceRoleNotificationService, userManager);
+            VerifyAllMocks(userResourceRoleNotificationService, globalRoleService, userPublicationRoleRepository);
         }
 
         [Fact]
@@ -912,8 +436,10 @@ public abstract class UserRoleServiceTests
                 .WithPublication(publication)
                 .WithRole(PublicationRole.Drafter);
 
-            var userManager = MockUserManager();
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(RoleNames.Analyst));
+            var globalRoleService = new Mock<IGlobalRoleService>(Strict);
+            globalRoleService
+                .Setup(mock => mock.UpgradeToGlobalRoleIfRequired(ItIsUser(user), RoleNames.Analyst))
+                .Returns(Task.CompletedTask);
 
             var userRepository = new Mock<IUserRepository>(Strict);
             userRepository
@@ -952,7 +478,7 @@ public abstract class UserRoleServiceTests
                     usersAndRolesDbContext: userAndRolesDbContext,
                     contentDbContext: contentDbContext,
                     userResourceRoleNotificationService: userResourceRoleNotificationService.Object,
-                    userManager: userManager.Object,
+                    globalRoleService: globalRoleService.Object,
                     userPublicationRoleRepository: userPublicationRoleRepository.Object,
                     userRepository: userRepository.Object
                 );
@@ -964,290 +490,7 @@ public abstract class UserRoleServiceTests
 
             VerifyAllMocks(
                 userResourceRoleNotificationService,
-                userManager,
-                userPublicationRoleRepository,
-                userRepository
-            );
-        }
-
-        [Fact]
-        public async Task ActiveUserExists_HasNoRoles_UpgradesRole()
-        {
-            var user = new ApplicationUser { Id = _user.Id.ToString(), Email = _user.Email };
-
-            Publication publication = _dataFixture.DefaultPublication();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                userAndRolesDbContext.Users.Add(user);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                contentDbContext.Publications.Add(publication);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            UserPublicationRole createdUserDrafterRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(_user)
-                .WithPublication(publication)
-                .WithRole(PublicationRole.Drafter);
-
-            var userManager = MockUserManager();
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(user))).ReturnsAsync([]);
-            userManager
-                .Setup(s => s.AddToRoleAsync(ItIsUser(user), RoleNames.Analyst))
-                .ReturnsAsync(new IdentityResult());
-
-            var userRepository = new Mock<IUserRepository>(Strict);
-            userRepository
-                .Setup(mock => mock.FindUserByEmail(_user.Email, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(_user);
-            userRepository
-                .Setup(mock => mock.FindActiveUserByEmail(_user.Email, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(_user);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, []);
-            userPublicationRoleRepository
-                .Setup(s =>
-                    s.Create(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Drafter,
-                        _user.Id,
-                        null,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(createdUserDrafterRole);
-
-            var userResourceRoleNotificationService = new Mock<IUserResourceRoleNotificationService>(Strict);
-            userResourceRoleNotificationService
-                .Setup(mock =>
-                    mock.NotifyUserOfNewDrafterRole(createdUserDrafterRole.Id, It.IsAny<CancellationToken>())
-                )
-                .Returns(Task.CompletedTask);
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    contentDbContext: contentDbContext,
-                    userResourceRoleNotificationService: userResourceRoleNotificationService.Object,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userRepository: userRepository.Object
-                );
-
-                var result = await service.InviteDrafter(_user.Email, publication.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(
-                userResourceRoleNotificationService,
-                userManager,
-                userPublicationRoleRepository,
-                userRepository
-            );
-        }
-
-        [Fact]
-        public async Task ActiveUserExists_HasHigherGlobalRoleThanAnalyst_DoesNotUpgradeRole()
-        {
-            var user = new ApplicationUser { Id = _user.Id.ToString(), Email = _user.Email };
-
-            Publication publication = _dataFixture.DefaultPublication();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                userAndRolesDbContext.Users.Add(user);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                contentDbContext.Publications.Add(publication);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            UserPublicationRole createdUserDrafterRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(_user)
-                .WithPublication(publication)
-                .WithRole(PublicationRole.Drafter);
-
-            var userManager = MockUserManager();
-            // Here we are testing that if the user already has a higher-powered role than Analyst, there's no need
-            // to assign them the lower-powered Analyst role.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(RoleNames.BauUser));
-
-            var userRepository = new Mock<IUserRepository>(Strict);
-            userRepository
-                .Setup(mock => mock.FindUserByEmail(_user.Email, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(_user);
-            userRepository
-                .Setup(mock => mock.FindActiveUserByEmail(_user.Email, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(_user);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, []);
-            userPublicationRoleRepository
-                .Setup(s =>
-                    s.Create(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Drafter,
-                        _user.Id,
-                        null,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(createdUserDrafterRole);
-
-            var userResourceRoleNotificationService = new Mock<IUserResourceRoleNotificationService>(Strict);
-            userResourceRoleNotificationService
-                .Setup(mock =>
-                    mock.NotifyUserOfNewDrafterRole(createdUserDrafterRole.Id, It.IsAny<CancellationToken>())
-                )
-                .Returns(Task.CompletedTask);
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    contentDbContext: contentDbContext,
-                    userResourceRoleNotificationService: userResourceRoleNotificationService.Object,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userRepository: userRepository.Object
-                );
-
-                var result = await service.InviteDrafter(_user.Email, publication.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(
-                userResourceRoleNotificationService,
-                userManager,
-                userPublicationRoleRepository,
-                userRepository
-            );
-
-            userManager.Verify(
-                mock => mock.AddToRoleAsync(It.IsAny<ApplicationUser>(), It.IsAny<string>()),
-                Times.Never
-            );
-
-            userManager.Verify(
-                mock => mock.RemoveFromRolesAsync(It.IsAny<ApplicationUser>(), It.IsAny<List<string>>()),
-                Times.Never
-            );
-        }
-
-        [Fact]
-        public async Task ActiveUserExists_HasLowerGlobalRoleThanAnalyst_UpgradesRole()
-        {
-            var user = new ApplicationUser { Id = _user.Id.ToString(), Email = _user.Email };
-
-            Publication publication = _dataFixture.DefaultPublication();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                userAndRolesDbContext.Users.Add(user);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            {
-                contentDbContext.Publications.Add(publication);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            UserPublicationRole createdUserDrafterRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(_user)
-                .WithPublication(publication)
-                .WithRole(PublicationRole.Drafter);
-
-            var userManager = MockUserManager();
-            // Here we are checking to see that the user is "upgraded" to the higher-powered Analyst role by their
-            // existing PreReleaseUser role being removed and the higher-powered Analyst role being added.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(RoleNames.PrereleaseUser));
-            userManager
-                .Setup(s => s.AddToRoleAsync(ItIsUser(user), RoleNames.Analyst))
-                .ReturnsAsync(new IdentityResult());
-            userManager
-                .Setup(s =>
-                    s.RemoveFromRolesAsync(ItIsUser(user), ItIs.ListSequenceEqualTo(ListOf(RoleNames.PrereleaseUser)))
-                )
-                .ReturnsAsync(new IdentityResult());
-
-            var userRepository = new Mock<IUserRepository>(Strict);
-            userRepository
-                .Setup(mock => mock.FindUserByEmail(_user.Email, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(_user);
-            userRepository
-                .Setup(mock => mock.FindActiveUserByEmail(_user.Email, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(_user);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, []);
-            userPublicationRoleRepository
-                .Setup(s =>
-                    s.Create(
-                        _user.Id,
-                        publication.Id,
-                        PublicationRole.Drafter,
-                        _user.Id,
-                        null,
-                        It.IsAny<CancellationToken>()
-                    )
-                )
-                .ReturnsAsync(createdUserDrafterRole);
-
-            var userResourceRoleNotificationService = new Mock<IUserResourceRoleNotificationService>(Strict);
-            userResourceRoleNotificationService
-                .Setup(mock =>
-                    mock.NotifyUserOfNewDrafterRole(createdUserDrafterRole.Id, It.IsAny<CancellationToken>())
-                )
-                .Returns(Task.CompletedTask);
-
-            await using (var contentDbContext = InMemoryApplicationDbContext(contentDbContextId))
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    contentDbContext: contentDbContext,
-                    userResourceRoleNotificationService: userResourceRoleNotificationService.Object,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userRepository: userRepository.Object
-                );
-
-                var result = await service.InviteDrafter(_user.Email, publication.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(
-                userResourceRoleNotificationService,
-                userManager,
+                globalRoleService,
                 userPublicationRoleRepository,
                 userRepository
             );
@@ -1495,214 +738,6 @@ public abstract class UserRoleServiceTests
 
             var result = await service.InviteDrafter("test@test.com", Guid.NewGuid());
 
-            result.AssertNotFound();
-        }
-    }
-
-    public class UpgradeToGlobalRoleIfRequiredTests : UserRoleServiceTests
-    {
-        [Fact]
-        public async Task UpgradeToGlobalRoleIfRequired_DoUpgrade()
-        {
-            var userId = Guid.NewGuid();
-            var user = new ApplicationUser { Id = userId.ToString() };
-
-            var existingRole = new IdentityRole { Id = Guid.NewGuid().ToString(), Name = RoleNames.PrereleaseUser };
-
-            var newRole = new IdentityRole { Id = Guid.NewGuid().ToString(), Name = RoleNames.Analyst };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddRangeAsync(user);
-                await userAndRolesDbContext.Roles.AddRangeAsync(existingRole, newRole);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // Here we are setting up the user so that they have a different Global Role currently assigned
-            // than the new one being set. They will have the existing one removed and the new one added.
-            userManager.Setup(mock => mock.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(existingRole.Name));
-
-            userManager
-                .Setup(mock => mock.AddToRoleAsync(ItIsUser(user), newRole.Name))
-                .ReturnsAsync(new IdentityResult());
-
-            userManager
-                .Setup(mock =>
-                    mock.RemoveFromRolesAsync(ItIsUser(user), ItIs.ListSequenceEqualTo(ListOf(existingRole.Name)))
-                )
-                .ReturnsAsync(new IdentityResult());
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object
-                );
-
-                await service.UpgradeToGlobalRoleIfRequired(RoleNames.Analyst, user);
-
-                VerifyAllMocks(userManager);
-            }
-        }
-
-        [Fact]
-        public async Task UpgradeToGlobalRoleIfRequired_DoNotUpgrade()
-        {
-            var userId = Guid.NewGuid();
-            var user = new ApplicationUser { Id = userId.ToString() };
-
-            var existingRole = new IdentityRole { Id = Guid.NewGuid().ToString(), Name = RoleNames.Analyst };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddRangeAsync(user);
-                await userAndRolesDbContext.Roles.AddRangeAsync(existingRole);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // user already has the correct role
-            userManager.Setup(mock => mock.GetRolesAsync(ItIsUser(user))).ReturnsAsync(ListOf(existingRole.Name));
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object
-                );
-
-                await service.UpgradeToGlobalRoleIfRequired(RoleNames.Analyst, user);
-
-                VerifyAllMocks(userManager);
-            }
-        }
-    }
-
-    public class GetAllGlobalRolesTests : UserRoleServiceTests
-    {
-        [Fact]
-        public async Task GetAllGlobalRoles()
-        {
-            var role1 = new IdentityRole
-            {
-                Id = Guid.NewGuid().ToString(),
-                Name = "Role 1",
-                NormalizedName = "ROLE 1",
-            };
-
-            var role2 = new IdentityRole
-            {
-                Id = Guid.NewGuid().ToString(),
-                Name = "Role 2",
-                NormalizedName = "ROLE 2",
-            };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddRangeAsync(role1, role2);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(usersAndRolesDbContext: userAndRolesDbContext);
-
-                var result = await service.GetAllGlobalRoles();
-
-                result.AssertRight();
-
-                var globalRoles = result.Right.ToList();
-                Assert.Equal(2, globalRoles.Count);
-
-                Assert.Equal(role1.Id, globalRoles[0].Id);
-                Assert.Equal(role1.Name, globalRoles[0].Name);
-                Assert.Equal(role1.NormalizedName, globalRoles[0].NormalizedName);
-
-                Assert.Equal(role2.Id, globalRoles[1].Id);
-                Assert.Equal(role2.Name, globalRoles[1].Name);
-                Assert.Equal(role2.NormalizedName, globalRoles[1].NormalizedName);
-            }
-        }
-    }
-
-    public class GetGlobalRolesForUserTests : UserRoleServiceTests
-    {
-        [Fact]
-        public async Task Success()
-        {
-            var user = new ApplicationUser { Id = Guid.NewGuid().ToString() };
-
-            var role1 = new IdentityRole
-            {
-                Id = Guid.NewGuid().ToString(),
-                Name = "Role 1",
-                NormalizedName = "ROLE 1",
-            };
-
-            var role2 = new IdentityRole
-            {
-                Id = Guid.NewGuid().ToString(),
-                Name = "Role 2",
-                NormalizedName = "ROLE 2",
-            };
-
-            // Role not assigned
-            var role3 = new IdentityRole
-            {
-                Id = Guid.NewGuid().ToString(),
-                Name = "Role 3",
-                NormalizedName = "ROLE 3",
-            };
-
-            var userRole1 = new IdentityUserRole<string> { UserId = user.Id, RoleId = role1.Id };
-
-            var userRole2 = new IdentityUserRole<string> { UserId = user.Id, RoleId = role2.Id };
-
-            var userAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.AddAsync(user);
-                await userAndRolesDbContext.AddRangeAsync(role1, role2, role3);
-                await userAndRolesDbContext.AddRangeAsync(userRole1, userRole2);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(userAndRolesDbContextId))
-            {
-                var service = SetupService(usersAndRolesDbContext: userAndRolesDbContext);
-
-                var result = await service.GetGlobalRolesForUser(user.Id);
-
-                result.AssertRight();
-
-                var globalRoles = result.Right;
-                Assert.Equal(2, globalRoles.Count);
-
-                Assert.Equal(role1.Id, globalRoles[0].Id);
-                Assert.Equal(role1.Name, globalRoles[0].Name);
-                Assert.Equal(role1.NormalizedName, globalRoles[0].NormalizedName);
-
-                Assert.Equal(role2.Id, globalRoles[1].Id);
-                Assert.Equal(role2.Name, globalRoles[1].Name);
-                Assert.Equal(role2.NormalizedName, globalRoles[1].NormalizedName);
-            }
-        }
-
-        [Fact]
-        public async Task NoUser_ReturnsNotFound()
-        {
-            await using var userAndRolesDbContext = InMemoryUserAndRolesDbContext();
-            var service = SetupService(usersAndRolesDbContext: userAndRolesDbContext);
-
-            var result = await service.GetGlobalRolesForUser(Guid.NewGuid().ToString());
             result.AssertNotFound();
         }
     }
@@ -1961,7 +996,7 @@ public abstract class UserRoleServiceTests
     public class RemoveUserPublicationRoleTests : UserRoleServiceTests
     {
         [Fact]
-        public async Task RemoveUserPublicationRole()
+        public async Task UserActive()
         {
             User user = _dataFixture.DefaultUser();
 
@@ -1981,16 +1016,13 @@ public abstract class UserRoleServiceTests
                 await userAndRolesDbContext.SaveChangesAsync();
             }
 
-            var userManager = MockUserManager();
+            var globalRoleService = new Mock<IGlobalRoleService>(Strict);
             var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
             var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
 
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(identityUser))).ReturnsAsync(ListOf(RoleNames.Analyst));
-            userManager
-                .Setup(s =>
-                    s.RemoveFromRolesAsync(ItIsUser(identityUser), ItIs.ListSequenceEqualTo(ListOf(RoleNames.Analyst)))
-                )
-                .ReturnsAsync(new IdentityResult());
+            globalRoleService
+                .Setup(mock => mock.DowngradeFromGlobalRoleIfRequired(ItIsUser(identityUser), RoleNames.Analyst))
+                .Returns(Task.CompletedTask);
 
             userPublicationRoleRepository
                 .Setup(m => m.GetById(userPublicationRole.Id, It.IsAny<CancellationToken>()))
@@ -2004,7 +1036,7 @@ public abstract class UserRoleServiceTests
             {
                 var service = SetupService(
                     usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object,
+                    globalRoleService: globalRoleService.Object,
                     userPublicationRoleRepository: userPublicationRoleRepository.Object,
                     userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
                 );
@@ -2014,7 +1046,44 @@ public abstract class UserRoleServiceTests
                 result.AssertRight();
             }
 
-            VerifyAllMocks(userManager, userPublicationRoleRepository);
+            VerifyAllMocks(globalRoleService, userPublicationRoleRepository);
+        }
+
+        [Theory]
+        [MemberData(nameof(AllTypesOfNonActiveUser))]
+        public async Task UserNotActive_DoesNotTryToChangeGlobalRole(Func<DataFixture, User> userFactory)
+        {
+            User user = userFactory(_dataFixture);
+
+            var identityUser = new ApplicationUser { Id = user.Id.ToString() };
+
+            UserPublicationRole userPublicationRole = _dataFixture
+                .DefaultUserPublicationRole()
+                .WithUser(user)
+                .WithPublication(_dataFixture.DefaultPublication())
+                .WithRole(PublicationRole.Drafter);
+
+            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
+            var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
+
+            userPublicationRoleRepository
+                .Setup(m => m.GetById(userPublicationRole.Id, It.IsAny<CancellationToken>()))
+                .ReturnsAsync(userPublicationRole);
+            userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
+            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
+
+            userPreReleaseRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
+
+            var service = SetupService(
+                userPublicationRoleRepository: userPublicationRoleRepository.Object,
+                userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
+            );
+
+            var result = await service.RemoveUserPublicationRole(userPublicationRole.Id);
+
+            result.AssertRight();
+
+            VerifyAllMocks(userPublicationRoleRepository);
         }
 
         [Fact]
@@ -2059,199 +1128,12 @@ public abstract class UserRoleServiceTests
 
             VerifyAllMocks(userPublicationRoleRepository);
         }
-
-        [Fact]
-        public async Task RemoveUserPublicationRole_HasHigherGlobalRoleThanAnalyst()
-        {
-            User user = _dataFixture.DefaultUser();
-
-            var identityUser = new ApplicationUser { Id = user.Id.ToString() };
-
-            UserPublicationRole userPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(user)
-                .WithPublication(_dataFixture.DefaultPublication())
-                .WithRole(PublicationRole.Drafter);
-
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddAsync(identityUser);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // Here the user has the BAU role, which is higher-powered than Analyst, so we will test that they remain
-            // a BAU user and do not have any roles removed.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(identityUser))).ReturnsAsync(ListOf(RoleNames.BauUser));
-
-            var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
-            userPreReleaseRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository
-                .Setup(m => m.GetById(userPublicationRole.Id, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(userPublicationRole);
-            userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
-                );
-
-                var result = await service.RemoveUserPublicationRole(userPublicationRole.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userManager, userPublicationRoleRepository, userPreReleaseRoleRepository);
-        }
-
-        [Fact]
-        public async Task RemoveUserPublicationRole_AnalystRoleStillRequiredForOtherPublications()
-        {
-            User user = _dataFixture.DefaultUser();
-
-            var identityUser = new ApplicationUser { Id = user.Id.ToString() };
-
-            UserPublicationRole userPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(user)
-                .WithPublication(_dataFixture.DefaultPublication())
-                .WithRole(PublicationRole.Drafter);
-
-            UserPublicationRole anotherUserPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(user)
-                .WithPublication(_dataFixture.DefaultPublication())
-                .WithRole(PublicationRole.Drafter);
-
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddAsync(identityUser);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // Here the user has the Analyst role currently.  We will test that they retain the Analyst role  because
-            // they still have need of it in other Publication roles.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(identityUser))).ReturnsAsync(ListOf(RoleNames.Analyst));
-
-            var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
-            userPreReleaseRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository
-                .Setup(m => m.GetById(userPublicationRole.Id, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(userPublicationRole);
-            userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, anotherUserPublicationRole);
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
-                );
-
-                var result = await service.RemoveUserPublicationRole(userPublicationRole.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userManager, userPublicationRoleRepository, userPreReleaseRoleRepository);
-        }
-
-        [Fact]
-        public async Task RemoveUserPublicationRole_DowngradeToPreReleaseRole()
-        {
-            User user = _dataFixture.DefaultUser();
-
-            var identityUser = new ApplicationUser { Id = user.Id.ToString() };
-
-            UserPublicationRole userPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(user)
-                .WithPublication(_dataFixture.DefaultPublication())
-                .WithRole(PublicationRole.Drafter);
-
-            UserReleaseRole preReleaseRole = _dataFixture
-                .DefaultUserPreReleaseRole()
-                .WithUser(user)
-                .WithReleaseVersion(
-                    _dataFixture
-                        .DefaultReleaseVersion()
-                        .WithRelease(_dataFixture.DefaultRelease().WithPublication(_dataFixture.DefaultPublication()))
-                );
-
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddAsync(identityUser);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // Here the user has the Analyst role currently.  We will test that they lose the Analyst role but gain the
-            // PreReleaseUser role, as they have need of this role elsewhere.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(identityUser))).ReturnsAsync(ListOf(RoleNames.Analyst));
-
-            userManager
-                .Setup(s =>
-                    s.RemoveFromRolesAsync(ItIsUser(identityUser), ItIs.ListSequenceEqualTo(ListOf(RoleNames.Analyst)))
-                )
-                .ReturnsAsync(new IdentityResult());
-
-            userManager
-                .Setup(s => s.AddToRoleAsync(ItIsUser(identityUser), RoleNames.PrereleaseUser))
-                .ReturnsAsync(new IdentityResult());
-
-            var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
-            userPreReleaseRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, preReleaseRole);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository
-                .Setup(m => m.GetById(userPublicationRole.Id, It.IsAny<CancellationToken>()))
-                .ReturnsAsync(userPublicationRole);
-            userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
-                );
-
-                var result = await service.RemoveUserPublicationRole(userPublicationRole.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userManager, userPublicationRoleRepository, userPreReleaseRoleRepository);
-        }
     }
 
     public class RemoveDrafterTests : UserRoleServiceTests
     {
         [Fact]
-        public async Task Success()
+        public async Task UserActive()
         {
             User user = _dataFixture.DefaultUser();
 
@@ -2271,16 +1153,13 @@ public abstract class UserRoleServiceTests
                 await userAndRolesDbContext.SaveChangesAsync();
             }
 
-            var userManager = MockUserManager();
+            var globalRoleService = new Mock<IGlobalRoleService>(Strict);
             var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
             var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
 
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(identityUser))).ReturnsAsync(ListOf(RoleNames.Analyst));
-            userManager
-                .Setup(s =>
-                    s.RemoveFromRolesAsync(ItIsUser(identityUser), ItIs.ListSequenceEqualTo(ListOf(RoleNames.Analyst)))
-                )
-                .ReturnsAsync(new IdentityResult());
+            globalRoleService
+                .Setup(mock => mock.DowngradeFromGlobalRoleIfRequired(ItIsUser(identityUser), RoleNames.Analyst))
+                .Returns(Task.CompletedTask);
 
             userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, userPublicationRole);
             userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
@@ -2292,7 +1171,7 @@ public abstract class UserRoleServiceTests
             {
                 var service = SetupService(
                     usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object,
+                    globalRoleService: globalRoleService.Object,
                     userPublicationRoleRepository: userPublicationRoleRepository.Object,
                     userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
                 );
@@ -2302,7 +1181,40 @@ public abstract class UserRoleServiceTests
                 result.AssertRight();
             }
 
-            VerifyAllMocks(userManager, userPublicationRoleRepository);
+            VerifyAllMocks(globalRoleService, userPublicationRoleRepository);
+        }
+
+        [Theory]
+        [MemberData(nameof(AllTypesOfNonActiveUser))]
+        public async Task UserNotActive_DoesNotTryToChangeGlobalRole(Func<DataFixture, User> userFactory)
+        {
+            User user = userFactory(_dataFixture);
+
+            UserPublicationRole userPublicationRole = _dataFixture
+                .DefaultUserPublicationRole()
+                .WithUser(user)
+                .WithPublication(_dataFixture.DefaultPublication())
+                .WithRole(PublicationRole.Drafter);
+
+            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
+            var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
+
+            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, userPublicationRole);
+            userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
+            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
+
+            userPreReleaseRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
+
+            var service = SetupService(
+                userPublicationRoleRepository: userPublicationRoleRepository.Object,
+                userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
+            );
+
+            var result = await service.RemoveDrafter(userPublicationRole.Id);
+
+            result.AssertRight();
+
+            VerifyAllMocks(userPublicationRoleRepository);
         }
 
         [Fact]
@@ -2363,187 +1275,6 @@ public abstract class UserRoleServiceTests
             );
 
             VerifyAllMocks(userPublicationRoleRepository);
-        }
-
-        [Fact]
-        public async Task HasHigherGlobalRoleThanAnalyst()
-        {
-            User user = _dataFixture.DefaultUser();
-
-            var identityUser = new ApplicationUser { Id = user.Id.ToString() };
-
-            UserPublicationRole userPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(user)
-                .WithPublication(_dataFixture.DefaultPublication())
-                .WithRole(PublicationRole.Drafter);
-
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddAsync(identityUser);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // Here the user has the BAU role, which is higher-powered than Analyst, so we will test that they remain
-            // a BAU user and do not have any roles removed.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(identityUser))).ReturnsAsync(ListOf(RoleNames.BauUser));
-
-            var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
-            userPreReleaseRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, userPublicationRole);
-            userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
-                );
-
-                var result = await service.RemoveDrafter(userPublicationRole.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userManager, userPublicationRoleRepository, userPreReleaseRoleRepository);
-        }
-
-        [Fact]
-        public async Task AnalystRoleStillRequiredForOtherPublications()
-        {
-            User user = _dataFixture.DefaultUser();
-
-            var identityUser = new ApplicationUser { Id = user.Id.ToString() };
-
-            UserPublicationRole userPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(user)
-                .WithPublication(_dataFixture.DefaultPublication())
-                .WithRole(PublicationRole.Drafter);
-
-            UserPublicationRole anotherUserPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(user)
-                .WithPublication(_dataFixture.DefaultPublication())
-                .WithRole(PublicationRole.Drafter);
-
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddAsync(identityUser);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // Here the user has the Analyst role currently.  We will test that they retain the Analyst role  because
-            // they still have need of it in other Publication roles.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(identityUser))).ReturnsAsync(ListOf(RoleNames.Analyst));
-
-            var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
-            userPreReleaseRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, userPublicationRole);
-            userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, anotherUserPublicationRole);
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
-                );
-
-                var result = await service.RemoveDrafter(userPublicationRole.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userManager, userPublicationRoleRepository, userPreReleaseRoleRepository);
-        }
-
-        [Fact]
-        public async Task DowngradeToPreReleaseRole()
-        {
-            User user = _dataFixture.DefaultUser();
-
-            var identityUser = new ApplicationUser { Id = user.Id.ToString() };
-
-            UserPublicationRole userPublicationRole = _dataFixture
-                .DefaultUserPublicationRole()
-                .WithUser(user)
-                .WithPublication(_dataFixture.DefaultPublication())
-                .WithRole(PublicationRole.Drafter);
-
-            UserReleaseRole preReleaseRole = _dataFixture
-                .DefaultUserPreReleaseRole()
-                .WithUser(user)
-                .WithReleaseVersion(
-                    _dataFixture
-                        .DefaultReleaseVersion()
-                        .WithRelease(_dataFixture.DefaultRelease().WithPublication(_dataFixture.DefaultPublication()))
-                );
-
-            var usersAndRolesDbContextId = Guid.NewGuid().ToString();
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                await userAndRolesDbContext.Users.AddAsync(identityUser);
-                await userAndRolesDbContext.SaveChangesAsync();
-            }
-
-            var userManager = MockUserManager();
-
-            // Here the user has the Analyst role currently.  We will test that they lose the Analyst role but gain the
-            // PreReleaseUser role, as they have need of this role elsewhere.
-            userManager.Setup(s => s.GetRolesAsync(ItIsUser(identityUser))).ReturnsAsync(ListOf(RoleNames.Analyst));
-
-            userManager
-                .Setup(s =>
-                    s.RemoveFromRolesAsync(ItIsUser(identityUser), ItIs.ListSequenceEqualTo(ListOf(RoleNames.Analyst)))
-                )
-                .ReturnsAsync(new IdentityResult());
-
-            userManager
-                .Setup(s => s.AddToRoleAsync(ItIsUser(identityUser), RoleNames.PrereleaseUser))
-                .ReturnsAsync(new IdentityResult());
-
-            var userPreReleaseRoleRepository = new Mock<IUserPreReleaseRoleRepository>(Strict);
-            userPreReleaseRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, preReleaseRole);
-
-            var userPublicationRoleRepository = new Mock<IUserPublicationRoleRepository>(Strict);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.All, userPublicationRole);
-            userPublicationRoleRepository.Setup(m => m.RemoveById(userPublicationRole.Id, default)).ReturnsAsync(true);
-            userPublicationRoleRepository.SetupQuery(ResourceRoleFilter.ActiveOnly, []);
-
-            await using (var userAndRolesDbContext = InMemoryUserAndRolesDbContext(usersAndRolesDbContextId))
-            {
-                var service = SetupService(
-                    usersAndRolesDbContext: userAndRolesDbContext,
-                    userManager: userManager.Object,
-                    userPublicationRoleRepository: userPublicationRoleRepository.Object,
-                    userPreReleaseRoleRepository: userPreReleaseRoleRepository.Object
-                );
-
-                var result = await service.RemoveDrafter(userPublicationRole.Id);
-
-                result.AssertRight();
-            }
-
-            VerifyAllMocks(userManager, userPublicationRoleRepository, userPreReleaseRoleRepository);
         }
     }
 
@@ -2628,7 +1359,8 @@ public abstract class UserRoleServiceTests
         IUserPreReleaseRoleRepository? userPreReleaseRoleRepository = null,
         IUserRepository? userRepository = null,
         UserManager<ApplicationUser>? userManager = null,
-        IUserService? userService = null
+        IUserService? userService = null,
+        IGlobalRoleService? globalRoleService = null
     )
     {
         contentDbContext ??= InMemoryApplicationDbContext();
@@ -2647,7 +1379,8 @@ public abstract class UserRoleServiceTests
             userPublicationRoleRepository ?? Mock.Of<IUserPublicationRoleRepository>(Strict),
             userPreReleaseRoleRepository ?? Mock.Of<IUserPreReleaseRoleRepository>(Strict),
             userRepository ?? Mock.Of<IUserRepository>(Strict),
-            userManager ?? MockUserManager().Object
+            userManager ?? MockUserManager().Object,
+            globalRoleService: Mock.Of<IGlobalRoleService>(Strict)
         );
     }
 }
