+++ src/explore-education-statistics-admin/src/pages/publication/components/InvitePublicationDrafterForm.tsx
@@ -57,6 +57,7 @@ const InvitePublicationDrafterForm = ({
         errorMappings={errorMappings}
         initialValues={inviteDrafterInitialValues}
         validationSchema={inviteDrafterValidationSchema}
+        reValidateMode="onSubmit"
         resetAfterSubmit
       >
         {({ formState }) => {
